refactor: migrate common types & enums into a single file
