+++ src/shell/expander/expander.ts
@@ -1,13 +1,11 @@
-import { Token, TokenType } from "../lexer/lexer.ts";
-import { AstNode, CommandNode, SimpleCommand } from "../parser/parser.ts";
-
-export type ExecutionCommandNode = ExecutionCommand;
-
-export interface ExecutionCommand extends AstNode {
-  readonly type: "ExecutionCommand";
-  readonly name: string;
-  readonly args: string[];
-}
+import {
+  Token,
+  CommandNode,
+  SimpleCommand,
+  ExecutionCommandNode,
+  ExecutionCommand,
+  TokenType,
+} from "../common.ts";
 
 export default class Expander {
   /**
