+++ src/shell/lexer/handler/handler.ts
@@ -1,4 +1,5 @@
-import Lexer, { Token } from "../lexer.ts";
+import Lexer from "../lexer.ts";
+import { Token } from "../../common.ts";
 
 export interface Handler {
   nextToken(lexer: Lexer, token: Token): void;
