+++ src/util/command_util.ts
@@ -1,13 +1,14 @@
 // @ts-expect-error eslint-disable-next-line @typescript-eslint/ban-ts-comment
 import getopts, { Options, ParsedOptions } from "getopts";
-import Expander, { ExecutionCommand } from "../shell/expander/expander.ts";
+import Expander from "../shell/expander/expander.ts";
 import { CommandScript } from "../command/command_script.ts";
 import TerminalUtil from "./terminal_util.ts";
 import CommandImportUtil from "./command_import_util.ts";
 import FileSystemUtil from "./file_system_util.ts";
 import { escape } from "lodash-es";
-import Parser, { SimpleCommand } from "../shell/parser/parser.ts";
+import Parser from "../shell/parser/parser.ts";
 import Lexer, { LexerError } from "../shell/lexer/lexer.ts";
+import { SimpleCommand, ExecutionCommand } from "../shell/common.ts";
 
 export default class CommandUtil {
   /**
