+++ test/unit/src/shell/lexer/handler/word_handler.test.ts
@@ -1,10 +1,10 @@
 import { beforeEach, describe, expect, test } from "vitest";
-import Lexer, {
-  LexerError,
+import Lexer, { LexerError } from "../../../../../../src/shell/lexer/lexer.ts";
+import {
   Token,
   TokenPartType,
   TokenType,
-} from "../../../../../../src/shell/lexer/lexer.ts";
+} from "../../../../../../src/shell/common.ts";
 import WordHandler from "../../../../../../src/shell/lexer/handler/word_handler.ts";
 
 describe("WordHandler", () => {
