+++ test/unit/src/shell/lexer/lexer.test.ts
@@ -1,10 +1,10 @@
 import { describe, expect, test, vi } from "vitest";
-import Lexer, {
-  LexerError,
+import Lexer, { LexerError } from "../../../../../src/shell/lexer/lexer.ts";
+import {
   Token,
   TokenPartType,
   TokenType,
-} from "../../../../../src/shell/lexer/lexer.ts";
+} from "../../../../../src/shell/common.ts";
 import { escape } from "lodash-es";
 import WordHandler from "../../../../../src/shell/lexer/handler/word_handler.ts";
